Move formatter into separate file
That way you can use it as pino formatter:
```
pino({
  prettyPrint: {
    formatter: require('pino-colada/formatter')()
  }
})
```
